🎨 Extract formatting to a new util
